Allow limiting the amount of subdomains to output in both active and passive modes.
Make a deepcopy of the Table object because otherwise we would only
have a reference to the Table object.
This is needed to allow the program to output a limited amount of
data to the CLI and the full amount to the file.

This was also done for the JSON data format as well.